+++ src/Utils/ruleForm.ts
@@ -106,7 +106,10 @@ export const schema = yup.object({
     name: 'price-not-allowed',
     message: 'Giá không phù hợp',
     test: testPriceMinMax
-  })
+  }),
+
+  //* rule cho feature search product
+  name: yup.string().trim().required('Tên sản phẩm là bắt buộc')
 })
 
 export type Schema = yup.InferType<typeof schema>
